test: split runtime review gate coverage
